--- Class1.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-public class Class1
-{
-	public Class1()
-	{
-
-		public int thing()
-        {
-			return 1;
-        }
-	}
-}
