--- Class2.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Calculator
-{
-    public class baseViewModel
-    {
-
-    }
-}
-
