+++ ToDoList/MainWindow.xaml.cs
@@ -23,33 +23,64 @@ namespace ToDoList
     /// </summary>
     public partial class MainWindow : Window
     {
-        private readonly string databaseName = "database.db";
 
-        private List<ToDoRecord> Records;
+        private ToDoListClass toDoList;
 
         public MainWindow()
         {
             InitializeComponent();
-            Records = new List<ToDoRecord>();
-            GetRecords();
+            toDoList = new ToDoListClass();
+            UpdateList();
         }
 
         private void saveButton_Click(object sender, RoutedEventArgs e)
         {
+            toDoList.SaveRecord(ToDoTitle, ToDoDetails);
+            UpdateList();
+        }
+
+        private void toDoEntries_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            var buff = sender as ListBox;
+            if (buff != null)
+            {
+                Debug.WriteLine(buff.SelectedItem);
+            }
+        }
+
+        private void UpdateList()
+        {
+            toDoList.GetRecords(toDoEntries);
+        }
+    }
+
+    public class ToDoListClass : baseViewModel
+    {
+        private readonly string databaseName = "database.db";
+        private List<ToDoRecord> Records;
+
+        public ToDoListClass()
+        {
+            Records = new List<ToDoRecord>();
+        }
+
+        public void SaveRecord(TextBox ToDoTitle, TextBox ToDoDetails)
+        {
             string title = ToDoTitle.Text;
             string desc = ToDoDetails.Text;
 
-            if (!string.IsNullOrWhiteSpace(title)) {
-                using (var connection = CommonFunctions.databaseOpenConnection(databaseName))
+            if (!string.IsNullOrWhiteSpace(title))
+            {
+                using (var connection = CommonFunctions.DatabaseOpenConnection(databaseName))
                 {
                     string insertString = string.Format("INSERT INTO test(Title, Desc) VALUES ('{0}', '{1}')", title, desc);
-                    var command = CommonFunctions.databaseCreateCommand(connection, insertString);
+                    var command = CommonFunctions.DatabaseCreateCommand(connection, insertString);
 
                     command.ExecuteNonQuery();
 
                     connection.Close();
                 }
-            } 
+            }
             else
             {
                 Debug.WriteLine("empty strings");
@@ -57,16 +88,14 @@ namespace ToDoList
 
             ToDoTitle.Text = string.Empty;
             ToDoDetails.Text = string.Empty;
-
-            GetRecords();
         }
 
-        private void GetRecords()
+        public void GetRecords(ListBox toDoEntries)
         {
             Records = new List<ToDoRecord>();
-            var connection = CommonFunctions.databaseOpenConnection(databaseName);
+            var connection = CommonFunctions.DatabaseOpenConnection(databaseName);
 
-            var command = CommonFunctions.databaseCreateCommand(connection, "SELECT * FROM test");
+            var command = CommonFunctions.DatabaseCreateCommand(connection, "SELECT * FROM test");
             var reader = command.ExecuteReader();
 
             while (reader.Read())
@@ -83,39 +112,11 @@ namespace ToDoList
             // Update list
             toDoEntries.Items.Clear();
 
-            foreach(var record in Records)
+            foreach (var record in Records)
             {
                 toDoEntries.Items.Add(record.ToString());
                 Debug.WriteLine(record.ToString());
             }
-        }
-
-        private void toDoEntries_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            var buff = sender as ListBox;
-            if (buff != null)
-            {
-                Debug.WriteLine(buff.SelectedItem);
-            }
-        }
-    }
-
-    public class ToDoRecord
-    {
-        public int Key;
-        public string Title;
-        public string Desc;
-
-        public ToDoRecord(int key, string title, string desc)
-        {
-            this.Key   = key;
-            this.Title = title;
-            this.Desc  = desc;
-        }
-
-        public override string ToString()
-        {
-            return string.Format("Key: {0}\nTitle: {1}\nDesc: {2}", Key, Title, Desc);
         }
     }
 }
